+++ src/UXI.GazeFilter/UXI.GazeFilter.csproj
@@ -72,6 +72,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BaseOptions.cs" />
+    <Compile Include="Extensions\FilterIOEx.cs" />
     <Compile Include="FileFormat.cs" />
     <Compile Include="FilterConfiguration.cs" />
     <Compile Include="FilterHost.cs" />
